+++ error_reporting.lua
@@ -9,7 +9,7 @@ local logger = named_logger.getLogger(NAME)
 local semver = require("semver")
 
 -- Module version
-local VERSION = semver(0, 8, 0)
+local VERSION = semver(0, 9, 0)
 
 -- Dependencies
 local serialization = require("serialization")